feat: New base middleware class provides utilities to traverse a model's ancestors looking for a particular condition to be met.
For example, User hasOne Company hasOne Product. To create a middleware handler that checks whether the caller is the owner of a particular product and has permission to edit it, use `AbstractBaseMiddleware.findAncestor(callerId, (model) => model === User)` to get the user who owns that product and then check whether it matches the caller.